feat: refactor contract-related imports and consolidate mock functions
