+++ packages/contract-deployment/src/index.ts
@@ -1,3 +1,6 @@
 export * from './deploy-contract';
+export * from './errors';
+export * from './find-deployed-contract';
 export * from './submit-deploy-tx';
+export * from './tx-model';
 export * from './unproven-deploy-tx';
