+++ packages/contract-governance/src/submit-replace-authority-tx.ts
@@ -14,13 +14,13 @@
  */
 
 import type { ContractAddress, SigningKey } from '@midnight-ntwrk/ledger';
+import { type ContractProviders } from '@midnight-ntwrk/midnight-js-contract-core';
+import { submitTx } from '@midnight-ntwrk/midnight-js-contract-core';
+import { createUnprovenReplaceAuthorityTx } from '@midnight-ntwrk/midnight-js-contract-core';
 import { type FinalizedTxData, SucceedEntirely } from '@midnight-ntwrk/midnight-js-types';
 import { assertDefined, assertIsContractAddress } from '@midnight-ntwrk/midnight-js-utils';
 
-import { type ContractProviders } from '@midnight-ntwrk/midnight-js-contract-core';
 import { ReplaceMaintenanceAuthorityTxFailedError } from './errors';
-import { submitTx } from './submit-tx';
-import { createUnprovenReplaceAuthorityTx } from './utils';
 
 /**
  * Constructs and submits a transaction that replaces the maintenance
