+++ packages/contract-mocks/src/index.ts
@@ -1,4 +1,1 @@
-export * from './mock-contracts';
 export * from './mock-providers';
-export * from './mock-transactions';
-export * from './test-helpers';
