--- packages/contract-mocks/src/mock-contracts.ts
@@ -1,6 +1,0 @@
-// Mock contracts for testing
-// TODO: Extract mock contracts from test files
-
-export const mockContract = {
-  // Placeholder - extract from actual tests
-};
