--- packages/contract-mocks/src/mock-transactions.ts
@@ -1,6 +1,0 @@
-// Mock transactions for testing
-// TODO: Extract mock transactions from test files
-
-export const mockTransaction = {
-  // Placeholder - extract from actual tests
-};
