--- packages/contract-mocks/src/test-helpers.ts
@@ -1,6 +1,0 @@
-// Test helpers and assertions
-// TODO: Add test helper utilities
-
-export const testHelpers = {
-  // Placeholder - add test helpers
-};
