+++ packages/contract-sdk/src/index.ts
@@ -1,6 +1,8 @@
 export * from './call';
 export * from './call-constructor';
-export * from './find-deployed-contract';
+export * from './errors';
 export * from './get-states';
 export * from './submit-call-tx';
+export * from './tx-interfaces';
+export * from './tx-model';
 export * from './unproven-call-tx';
