+++ packages/contracts/package.json
@@ -24,9 +24,10 @@
     "deploy": "yarn npm publish"
   },
   "dependencies": {
-    "@midnight-ntwrk/midnight-js-network-id": "workspace:*",
-    "@midnight-ntwrk/midnight-js-types": "workspace:*",
-    "@midnight-ntwrk/midnight-js-utils": "workspace:*"
+    "@midnight-ntwrk/midnight-js-contract-core": "workspace:*",
+    "@midnight-ntwrk/midnight-js-contract-deployment": "workspace:*",
+    "@midnight-ntwrk/midnight-js-contract-governance": "workspace:*",
+    "@midnight-ntwrk/midnight-js-contract-sdk": "workspace:*"
   },
   "files": [
     "dist/"
