+++ packages/contracts/src/index.ts
@@ -1,0 +1,4 @@
+export * from '@midnight-ntwrk/midnight-js-contract-core';
+export * from '@midnight-ntwrk/midnight-js-contract-deployment';
+export * from '@midnight-ntwrk/midnight-js-contract-governance';
+export * from '@midnight-ntwrk/midnight-js-contract-sdk';
